Move WebPage to marketplace.py
