+++ src/ai_marketplace_monitor/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = """Bo Peng"""
 __email__ = "ben.bob@gmail.com"
-__version__ = "0.8.6b"
+__version__ = "0.8.6b1"
