+++ tests/test_ai_marketplace_monitor.py
@@ -13,7 +13,7 @@ from ai_marketplace_monitor.user import User
 
 def test_version(version: str) -> None:
     """Sample pytest test function with the pytest fixture as an argument."""
-    assert version == "0.8.6b"
+    assert version == "0.8.6b1"
 
 
 def test_listing_cache(temp_cache: Cache, listing: Listing) -> None:
